+++ nssock/nssock.c
@@ -129,7 +129,7 @@ Ns_ModuleInit(const char *server, const char *module)
  */
 
 static NS_SOCKET
-SockListen(Ns_Driver *driver, const char *address, int port, int backlog)
+SockListen(Ns_Driver *driver, const char *address, unsigned short port, int backlog)
 {
     NS_SOCKET sock;
